refactor: 代码健康清理（死参数/Arc 共享/错误结构化/词汇单源）
审查修复批三前半：删 RuleSet::load 的 _engine_label 与
pipe_to_shell 的 commands 死参数（连带修正 main.rs engine_label 失
实注释）；FoundLayers.knowledge 改 Arc<KnowledgeBase> 一次性共享，
消除装配路径的整库深克隆；RuleSet::load 错误由 String 结构化为
RuleSetError（ExplicitConfig/Config/Script 三类，Display 保持原告
警文案）；allow/confirm/deny 词汇解析收敛 model::Decision::parse
单源（协议与 schema 反序列化共用）；log 降级指纹改 serve 实例局部
变量（消除跨实例污染）；endpoint_name canonicalize 失败 warn-once
告警；hook_decide 端点名算一次下传（重试环免重复 canonicalize）；
channel::emit 返回收窄 u8；清函数内重复 use。

+++ src/channel.rs
@@ -110,7 +110,7 @@ pub fn read_hook_input(agent: Agent) -> Option<HookInput> {
     })
 }
 
-pub fn emit(verdict: &Verdict, agent: Agent) -> i32 {
+pub fn emit(verdict: &Verdict, agent: Agent) -> u8 {
     let hook_envelope = |decision: &str| {
         println!(
             "{{\"hookSpecificOutput\":{{\"hookEventName\":\"PreToolUse\",\"permissionDecision\":\"{decision}\"}}}}"
